test(kmarshal): error serialization behavior
