+++ packages/kmarshal/package.json
@@ -26,6 +26,8 @@
     "@endo/marshal": "^1.8.0"
   },
   "devDependencies": {
+    "@endo/init": "^1.1.12",
+    "@endo/ses-ava": "^1.3.2",
     "ava": "^5.3.0"
   },
   "publishConfig": {
@@ -34,9 +36,6 @@
   "ava": {
     "files": [
       "test/**/*.test.*"
-    ],
-    "require": [
-      "@endo/init/debug.js"
     ],
     "timeout": "2m"
   },
